Optimize ASR folder structure
