+++ plugins/inputs/asr/asr_common.go
@@ -1,4 +1,4 @@
-package inputs
+package asr
 
 import (
 	"context"
