+++ plugins/inputs/asr/asr_common_test.go
@@ -1,4 +1,4 @@
-package inputs
+package asr
 
 import (
 	"testing"
