+++ plugins/inputs/asr/google_asr.go
@@ -1,4 +1,4 @@
-package inputs
+package asr
 
 import (
 	"context"
