+++ plugins/inputs/asr/google_asr_rtsp.go
@@ -1,4 +1,4 @@
-package inputs
+package asr
 
 import (
 	"context"
