+++ plugins/inputs/asr/parallel_asr_test.go
@@ -1,4 +1,4 @@
-package inputs
+package asr
 
 import (
 	"testing"
