+++ plugins/inputs/asr/riva_asr_test.go
@@ -1,4 +1,4 @@
-package inputs
+package asr
 
 import (
 	"encoding/binary"
