+++ plugins/inputs/inputs.go
@@ -1,0 +1,5 @@
+package inputs
+
+import (
+	_ "github.com/openmind/om1/plugins/inputs/asr"
+)
